pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-local-sage"
version = "0.1.0"
description = "AI-powered local tools for Obsidian - semantic search, document jump, link suggestions, and more"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your-email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your-email@example.com"}
]
keywords = [
    "obsidian",
    "ai",
    "semantic-search",
    "local-llm",
    "ollama",
    "qdrant",
    "vector-database",
    "knowledge-management",
    "note-taking",
    "productivity"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: General",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
bm25 = [
    "rank-bm25>=0.2.2"
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.28.0"
]
all = [
    "obsidian-local-sage[bm25,dev]"
]

[project.urls]
Homepage = "https://github.com/your-username/obsidian-local-sage"
Documentation = "https://github.com/your-username/obsidian-local-sage#readme"
Repository = "https://github.com/your-username/obsidian-local-sage.git"
Issues = "https://github.com/your-username/obsidian-local-sage/issues"
Changelog = "https://github.com/your-username/obsidian-local-sage/blob/main/CHANGELOG.md"

[project.scripts]
obsidian-sage = "src.cli:cli_main"
sage-search = "src.cli:cli_search"
sage-jump = "src.cli:cli_jump"
sage-ingest = "src.cli:cli_ingest"
sage-health = "src.cli:cli_health"

[tool.setuptools]
packages = ["src", "src.core", "src.search", "src.jump", "src.batch", "src.health", "src.analysis", "src.automation", "src.chat"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.sh"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | build
    | dist
    | __pycache__
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]
